+++ src/Combinations/repositories/Truncate.go
@@ -1,10 +1,10 @@
-package CombinationUtils
+package CombinationRepo
 
 import (
 	"context"
 	"errors"
 
-	CombinationGenerationCounterUtils "github.com/glener10/rotating-pairs-back/src/CombinationGenerationCounter/utils"
+	CombinationGenerationCounterRepo "github.com/glener10/rotating-pairs-back/src/CombinationGenerationCounter/repositories"
 	CommonRepository "github.com/glener10/rotating-pairs-back/src/common/repositories"
 	"go.mongodb.org/mongo-driver/bson"
 )
@@ -29,7 +29,7 @@ func CleanCollection() error {
 	if err != nil {
 		return errors.New("Error to DeleteMany in 'Combinations' collection:" + err.Error())
 	}
-	err = CombinationGenerationCounterUtils.Truncate()
+	err = CombinationGenerationCounterRepo.Truncate()
 	if err != nil {
 		return errors.New("Error to DeleteMany in 'TotalCombinationGenerationAccordingNumberOfInputs' collection:" + err.Error())
 	}