+++ src/Combinations/useCases/Combination.useCase.go
@@ -6,7 +6,7 @@ import (
 	"github.com/gin-gonic/gin"
 	CombinationGenerationCounterRepo "github.com/glener10/rotating-pairs-back/src/CombinationGenerationCounter/repositories"
 	CombinationRepo "github.com/glener10/rotating-pairs-back/src/Combinations/repositories"
-	CombinationRequestDto "github.com/glener10/rotating-pairs-back/src/common/interfaces"
+	CombinationRequestDto "github.com/glener10/rotating-pairs-back/src/common/dtos"
 )
 
 func Combination(c *gin.Context, combinationRequest CombinationRequestDto.CombinationRequest) {